+++ sqlalchemy_redshift/dialect.py
@@ -3,7 +3,7 @@ from collections import defaultdict, namedtuple
 
 import pkg_resources
 import sqlalchemy as sa
-from sqlalchemy import Column, exc, inspect
+from sqlalchemy import inspect
 from sqlalchemy.dialects.postgresql.base import (
     PGCompiler, PGDDLCompiler, PGIdentifierPreparer, PGTypeCompiler
 )
@@ -20,9 +20,11 @@ from sqlalchemy.dialects.postgresql import DOUBLE_PRECISION
 
 from .commands import (
     CopyCommand, UnloadFromSelect, Format, Compression, Encoding,
-    CreateLibraryCommand, AlterTableAppendCommand,
+    CreateLibraryCommand, AlterTableAppendCommand, RefreshMaterializedView
 )
-from .compat import string_types
+from .ddl import (
+    CreateMaterializedView, DropMaterializedView, get_table_attributes
+)
 
 try:
     import alembic
@@ -60,6 +62,9 @@ __all__ = (
 
     'CopyCommand', 'UnloadFromSelect', 'RedshiftDialect', 'Compression',
     'Encoding', 'Format', 'CreateLibraryCommand', 'AlterTableAppendCommand',
+    'RefreshMaterializedView',
+
+    'CreateMaterializedView', 'DropMaterializedView'
 )
 
 
@@ -330,42 +335,10 @@ class RedshiftDDLCompiler(PGDDLCompiler):
     """
 
     def post_create_table(self, table):
-        text = ""
+        kwargs = ["diststyle", "distkey", "sortkey", "interleaved_sortkey"]
         info = table.dialect_options['redshift']
-
-        diststyle = info.get('diststyle')
-        if diststyle:
-            diststyle = diststyle.upper()
-            if diststyle not in ('EVEN', 'KEY', 'ALL'):
-                raise exc.CompileError(
-                    u"diststyle {0} is invalid".format(diststyle)
-                )
-            text += " DISTSTYLE " + diststyle
-
-        distkey = info.get('distkey')
-        if distkey:
-            text += " DISTKEY ({0})".format(self.preparer.quote(distkey))
-
-        sortkey = info.get('sortkey')
-        interleaved_sortkey = info.get('interleaved_sortkey')
-        if sortkey and interleaved_sortkey:
-            raise exc.ArgumentError(
-                "Parameters sortkey and interleaved_sortkey are "
-                "mutually exclusive; you may not specify both."
-            )
-        if sortkey or interleaved_sortkey:
-            if isinstance(sortkey, string_types):
-                keys = [sortkey]
-            else:
-                keys = sortkey or interleaved_sortkey
-            keys = [key.name if isinstance(key, Column) else key
-                    for key in keys]
-            if interleaved_sortkey:
-                text += " INTERLEAVED"
-            sortkey_string = ", ".join(self.preparer.quote(key)
-                                       for key in keys)
-            text += " SORTKEY ({0})".format(sortkey_string)
-        return text
+        info = {key: info.get(key) for key in kwargs}
+        return get_table_attributes(self.preparer, **info)
 
     def get_column_specification(self, column, **kwargs):
         colspec = self.preparer.format_column(column)